Add y-axis to XY charts
The y-axis is displayed on the left of the chart, not obstructing the
data, and with the same background color of the chart.

Charts that do not have a y-axis, like the Resources Status chart, have
that space occupied by the component that exists on the left, like the
list of processes.

This way, all charts and the top horizontal rule remain vertically aligned.

Also dynamically changes the numbers on the axis, so that 1,000 appears
as 1K, 50,000,000 as 50M and so on.

Fixes #421.

Signed-off-by: Rodrigo Pinto <rodrigo.pinto@calian.ca>