Enable tooltips for scatter plots.
These tooltips behave as in the line charts: when the mouse stops moving for a
couple of seconds the tooltip is triggered.

If there are points close to the mouse pointer, the tooltip displays a
list with the closest point of each series. If the points are too far, the
tooltip is not shown.

The tooltip displays the timestamp that corresponds to the position of
the mouse, and y-value of the closest points.

This commit also enables the scatter plots, since the two issues that
were blocking their release (bug about lack of precision and the absence
of tooltip) have been resolved.

Signed-off-by: Rodrigo Pinto <rodrigo.pinto@calian.ca>