Add initial support for the Trace Overview view
Users of the trace extension would like to have an overview
of the trace that they are analyzing so they can select a range
that’s interesting, then filter, zoom, and get more details.
It also helps users maintain the context.

This commit adds initial support for the Trace Overview view
as part of the Histogram view. It puts the common code between
the trace-overview-component and xy-output-component in an abstract
class. Then, the trace-overview-component extends this class
to define custom behaviours. In the future, the xy-output-component
should also extend this abstract class. The behaviour of the
trace-overview-component should mimic the trace overview
of the Histogram view in the Trace Compass desktop application.

Signed-off-by: Hoang Thuan Pham <hoang.pham@calian.ca>