Fix css positioning for filter text box icon
Currently, the filter icon of the filter text box stays fixed
at the top left of the filter tree, even when the filter text
box is out of sight when users scroll down the tree. This commit
fixes this issue by setting the position css property of the
parent div of the filter text box to relative, so that the filter
icon is positioned relatively to the parent container and stays
attached to the filter text box.

Signed-off-by: Hoang Thuan Pham <hoang.pham@calian.ca>