Fix for title bar on Safari browsers
Titles of elements like charts, when too long, are cropped and display
an ellipsis at the end of the text. On Safari browsers this ellipsis is
not appearing.

Safari's `text-overflow: ellipsis` property works correctly, but when
used in conjunction with `writing-mode: vertical-rl`, apparently it
fails to display the ellipsis.

Fixes #764.

Signed-off-by: Rodrigo Pinto <rodrigo.pinto@calian.ca>