Do a 'yarn upgrade' before otherwise broken uplift
This upgrade is necessary so the following Theia version uplift still
works with the electron yarn commands.

Cap the fontawesome-svg-core version to less than 1.3.0. Without this
change, the hereby upgrade leads to a broken yarn build. This is despite
requesting for a version that remains compatible with 1.2.17. So a yarn
upgrade without this adjustment lends the use of v1.3.0, which breaks the
build. Below [1] flags that breaking version as deprecated, and the
later versions are way more major uplifts (6.x.y).

[1]https://www.npmjs.com/package/@fortawesome/fontawesome-svg-core?activeTab=versions

Signed-off-by: Marco Miller <marco.miller@ericsson.com>