Time Range Data Map
Isolates the code that keeps track of all opened experiment's / trace's time range data.
This code originally lived inside of the TimeRangeDataWidget react component, but the
logic is now isolated in it's own class so it can be used elsewhere without code duplication.

Adds a public method in the TimeRangeDataWidget where state and time range data can
be restored externally.

Signed-off-by: Will Yang<william.yang@ericsson.com>