Hide empty rows when filter is applied
Hide rows that have no non-null state in their model, when a filter is
applied. Gaps between states in the row model should be make rows
visible when strategy is SAMPLED, but not when it is DEEP.

Keep an up-to-date list of empty nodes after fetching row models from
the server.

Add emptyNodes and hideEmptyNodes to the properties of EntryTree, Tree
and Table. Filter the empty nodes when rendering the Table.

Removed unused dataRows in TimegraphOuputState

Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>