Improve display of vertical tree gridlines in timegraph views
v0.4.2 introduced vertical gridlines for the trees of timegraph views.
However, 2 vertical lines were drawn so that they appeared thicker. This
commit fixes that.

Besides, the vertical line on the right of the last column are not
drawn anymore. For that, the hideFiller flag needed to be added to the
EntryTree props so that that flag can be set to true, propagated and
hence the filler can be hidden.

Signed-off-by: Bernd Hufmann <bernd.hufmann@ericsson.com>