Enable sorting of timegraph tree columns
Change the sortable property to true for all defined columns in a
timegraph. Add a sortable default column with empty header when no
column headers are defined in the model.

Change the FilterTree onOrderChange callback to include all ordered ids,
including collapsed and hidden nodes.

Add a handler for the onOrderChange callback for a timegraph. When the
callback is handled, reorder the timegraphTree entries according to the
specified order. Update the timegraphTree in the state to trigger
synchronization of the timegraph chart.

Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>