[customized views] remove UIs related to managing them when irrelevant
In the Theia trace extension, the functionality to create
custom views has not been implemented. However the UI element
, a plus sign button added after the "views" entry of the
"configurator" element, used to trigger creating custom views,
is always present, even when not supported (no callback
provided that create the custom views). This commit removes
that button when not relevant.

There's also the UI element added to remove previously created
custom views (the "x" button), that does work in the Theia
extension. For consistency, this commit also removes this
button.

Another tweak done when customization is not supported is to
filter root "configurator" nodes from the nodes list early, so
they are not displayed in the views list. If they have children
, from previously configured output descriptors (from the
vscode trace extension), then the children will be displayed as
before but as root nodes instead of children of their
"configurator" node.

The filtering of "configurator" node is overridable, by
creating a class that extends ReactAvailableViewsWidget and
overrides method :
"protected doFilterList(list: OutputDescriptor[]): OutputDescriptor[]",

e.g. to return the original, unfiltered list.

Signed-off-by: Marc Dumais <marc.dumais@ericsson.com>