prevent traceviewer-* libraries' versions being updated by lerna
It would be preferable to manage the version of traceviewer base
and react libraries elsewhere. Lerna is stepping the various
packages versions, when we do releases, using the package.json
scripts entries, e.g. "yarn version:patch".

The trick used is to have lerna only step the version of public
packages (vscode-trace-extension only) and leave the others alone,
including the traceviewer-* libraries. For this to work, it was
necessary to mark the libraries as private - this change will have
to be undone wherever they might get published to npm.

Signed-off-by: Marc Dumais <marc.dumais@ericsson.com>